move HotRestart[Nop]Impl to Envoy::Server
