+++ test/integration/BUILD
@@ -130,7 +130,6 @@ envoy_cc_test_library(
         "//source/common/common:assert_lib",
         "//source/common/common:logger_lib",
         "//source/common/common:thread_lib",
-        "//source/common/hot_restart:hot_restart_nop_lib",
         "//source/common/http:codec_client_lib",
         "//source/common/http:header_map_lib",
         "//source/common/http:headers_lib",
@@ -145,6 +144,7 @@ envoy_cc_test_library(
         "//source/common/upstream:upstream_includes",
         "//source/common/upstream:upstream_lib",
         "//source/server:connection_handler_lib",
+        "//source/server:hot_restart_nop_lib",
         "//source/server:server_lib",
         "//source/server:test_hooks_lib",
         "//source/server/config/http:buffer_lib",
